requirements.txt
# ── Core AI Framework ─────────────────────────────────────────────────────────
crewai>=0.80.0
crewai-tools>=0.14.0
openai>=1.35.0

# ── Web Search (duckduckgo_search was renamed to ddgs) ────────────────────────
ddgs>=9.0.0

# ── Frontend ──────────────────────────────────────────────────────────────────
streamlit>=1.35.0

# ── Configuration & Validation ────────────────────────────────────────────────
pydantic>=2.7.0
pydantic-settings>=2.3.0
python-dotenv>=1.0.1

# ── Utilities ─────────────────────────────────────────────────────────────────
requests>=2.32.0
